updated scene 1
I have updated scene 1 so we have a hidden button on the right to open the wiring panel and also on the left to show a guide/cypher to explain to the user what wire they need to cut to escape this room. The right button press will show the user an image of wires and on the left button it will show the guide/cypher which will be an image which makes the user have to work out which wire they need to cut.

I have added a border to the game  which allows for a better view of the game and also a more "game" like feel.

I have added a generic time placeholder in the top left for when we eventually add a timer feature.

We still need to code the wires being cut and this being the win condition as well as making for the left button. We need to make the cypher to create some difficulty to this and make the user have to figure out which button is to be pressed.

I have coded a go home button in the top right so the user can press this to return back home.

I have coded a close button so when the user clicks on an image/hidden button, they can close it using the close button. We may need to move the close button slightly more in line with the game screen.

We also need to actually code the wire image so the wires can be clicked on and cut but only after clicking on something else first for example the hidden button on the left to show the guide so that the player cant just brute force their way through the puzzle as soon as they get into the room.