Changes to Scene 1
10/12/2024 - Scene 1 

I think this is mostly finished now, just a couple touch ups needed.

I've added a red outline to all buttons just whilst this is still being coded and finalized so we can see the exact size of each clickable button.

added a button for the door.
when initially clicked it tells the player to explore the area further.
when explored and then clicked it tells the player its locked and they need to figure out how to disable it.
when the correct wire is cut and it is clicked it will let the player know they can leave the area and it will send the player to the next area/level.

moved the "close" button slightly when the images are open for easier viewing.

added buttons for each wire and correctly adjusted them to be in-line with each wire respectively.

added themed symbols for each wire which the player will use to figure out which wire to cut as per the cypher.

added an image for the hidden button on the left.
this shows an image of a cypher the player needs to de-cypher to figure out which wire to cut.
mathematical equation will lead the user to a number which is assigned to a symbol. The symbol which it is assigned to is the symbol of wire the user needs to cut.
adjusted the positioning and size of the hidden left button.

added a condition on the wire buttons which checks if the player has clicked the hidden left button to view the cypher before cutting any wires. if not it will not let the player cut any wires.

adjusted the code for the buttons so they always appear over the top of images